Create Avg_func.m
Average function used for statistical analysis.  This function takes the accuracy and divergence of an imputed field with any missing vector filled in with the average known vector from the field.  It does this over the three test fields and takes the average accuracy and divergence from this average vector method over the three fields.